+++ utils/test-utils/src/main/scala/pl/touk/nussknacker/test/ExtremelyPatientScalaFutures.scala
@@ -1,9 +1,8 @@
 package pl.touk.nussknacker.test
 
-import org.scalatest.concurrent.{Eventually, ScalaFutures}
 import org.scalatest.time.{Millis, Minutes, Span}
 
-trait ExtremelyPatientScalaFutures extends ScalaFutures with Eventually {
+trait ExtremelyPatientScalaFutures extends BasePatientScalaFutures {
 
   final override implicit def patienceConfig: PatienceConfig = PatienceConfig(timeout = scaled(Span(5, Minutes)), interval = scaled(Span(100, Millis)))