+++ docs/Changelog.md
@@ -35,6 +35,7 @@
 * [#4131](https://github.com/TouK/nussknacker/pull/4131) Support for components using other languages than SpEL, added basic support for SpEL in template mode
 * [#4132](https://github.com/TouK/nussknacker/pull/4132) Combine statuses Failing, Failed, Error, Warning, FailedToGet and MulipleJobsRunning into one status that represents a "Problem".
   Status "Unknown" is removed.
+* [#4135](https://github.com/TouK/nussknacker/pull/4135) Added timeout configuration for fetching scenario state
 
 
 1.8.1 (28 Feb 2023)